*: Rely on eircode service only for address resolution
There was a lot of overlap in how the address was resolved. Ultimately,
the eircode service is a better candidate for keeping all this data as
it can also be used by other services.

The rough prefix resolver was removed and moved to the eircode service
Only one address resolver is used - eircode service

Because we need to authenticate with the service we have also added the
authentication token for the service.

Change-Id: Ib795879bdb5690f163cda84847a9da20443ab1df