SurrealDB Support (#206)
Since we are working on a Helm Charts, we have to add a non-sqlite
database support. After battling with trying to pull MySQL and
PostgreSQL from ArtfactHub, I decided that the easiest way would be to
implement SurrealDB support since we wanted one anyways.

The ID format had to slightly change to be using understores instead of
dashes, but it shouldn't affect anything else.

Also included are small bumps in the dependencies needed for this
change, and deprecation of the radius methods. They will later be
removed from the API but for now they return deprecation notice early.